AGENT_CONVERSATIONEL/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.118.0
fastapi-utils==0.8.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.23
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.11
mypy_extensions==1.1.0
ndjson==0.3.1
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.109.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
parso==0.8.5
peewee==3.17.8
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==5.9.8
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
pytz==2024.2
# pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.5.2
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.11.0
tokenizers==0.22.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xai-sdk==1.1.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
slowapi==0.1.9
